Move geo widget classes into widgets.geo package
