+++ collect_app/src/main/java/org/odk/collect/android/activities/FormFillingActivity.java
@@ -162,8 +162,8 @@ import org.odk.collect.android.utilities.InstancesRepositoryProvider;
 import org.odk.collect.android.utilities.MediaUtils;
 import org.odk.collect.android.utilities.SavepointsRepositoryProvider;
 import org.odk.collect.android.utilities.SoftKeyboardController;
-import org.odk.collect.android.widgets.GeoShapeWidget;
-import org.odk.collect.android.widgets.GeoTraceWidget;
+import org.odk.collect.android.widgets.geo.GeoShapeWidget;
+import org.odk.collect.android.widgets.geo.GeoTraceWidget;
 import org.odk.collect.android.widgets.MediaWidgetAnswerViewModel;
 import org.odk.collect.android.widgets.QuestionWidget;
 import org.odk.collect.android.widgets.datetime.DateTimeWidget;
