+++ collect_app/src/main/java/org/odk/collect/android/widgets/WidgetFactory.java
@@ -37,6 +37,9 @@ import org.odk.collect.android.widgets.barcode.BarcodeWidget;
 import org.odk.collect.android.widgets.datetime.DateTimeWidget;
 import org.odk.collect.android.widgets.datetime.DateWidget;
 import org.odk.collect.android.widgets.datetime.TimeWidget;
+import org.odk.collect.android.widgets.geo.GeoPointWidget;
+import org.odk.collect.android.widgets.geo.GeoShapeWidget;
+import org.odk.collect.android.widgets.geo.GeoTraceWidget;
 import org.odk.collect.android.widgets.image.ImageWidget;
 import org.odk.collect.android.widgets.items.LabelWidget;
 import org.odk.collect.android.widgets.items.LikertWidget;
