+++ collect_app/src/main/java/org/odk/collect/android/widgets/geo/GeoPointWidgetContent.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.widgets
+package org.odk.collect.android.widgets.geo
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -10,6 +10,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import org.javarosa.form.api.FormEntryPrompt
 import org.odk.collect.android.utilities.Appearances
+import org.odk.collect.android.widgets.MediaWidgetAnswerViewModel
+import org.odk.collect.android.widgets.WidgetAnswer
+import org.odk.collect.android.widgets.WidgetIconButton
 import org.odk.collect.androidshared.ui.compose.marginStandard
 import org.odk.collect.geo.GeoUtils
 import org.odk.collect.strings.R.string
