+++ collect_app/src/main/java/org/odk/collect/android/widgets/geo/GeoShapeWidget.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (C) 2015 GeoODK
- *
+ * Copyright (C) 2014 GeoODK
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
  *
@@ -12,7 +11,7 @@
  * the License.
  */
 
-package org.odk.collect.android.widgets;
+package org.odk.collect.android.widgets.geo;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
@@ -22,22 +21,20 @@ import android.view.View;
 
 import org.javarosa.core.model.data.IAnswerData;
 import org.javarosa.form.api.FormEntryPrompt;
-import org.odk.collect.android.databinding.GeotraceQuestionBinding;
+import org.odk.collect.android.databinding.GeoshapeQuestionBinding;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
+import org.odk.collect.android.widgets.QuestionWidget;
 import org.odk.collect.android.widgets.interfaces.GeoDataRequester;
 import org.odk.collect.android.widgets.utilities.GeoWidgetUtils;
 
-/**
- * GeoTraceWidget allows the user to collect a trace of GPS points as the
- * device moves along a path.
- */
 @SuppressLint("ViewConstructor")
-public class GeoTraceWidget extends QuestionWidget {
-    GeotraceQuestionBinding binding;
+public class GeoShapeWidget extends QuestionWidget {
+    GeoshapeQuestionBinding binding;
 
     private final GeoDataRequester geoDataRequester;
 
-    public GeoTraceWidget(Context context, QuestionDetails questionDetails, GeoDataRequester geoDataRequester, Dependencies dependencies) {
+    public GeoShapeWidget(Context context, QuestionDetails questionDetails,
+                          GeoDataRequester geoDataRequester, Dependencies dependencies) {
         super(context, dependencies, questionDetails);
         render();
 
@@ -46,13 +43,11 @@ public class GeoTraceWidget extends QuestionWidget {
 
     @Override
     protected View onCreateWidgetView(Context context, FormEntryPrompt prompt, int answerFontSize) {
-        binding = GeotraceQuestionBinding.inflate(((Activity) context).getLayoutInflater());
+        binding = GeoshapeQuestionBinding.inflate(((Activity) context).getLayoutInflater());
 
         binding.geoAnswerText.setTextSize(TypedValue.COMPLEX_UNIT_DIP, answerFontSize);
 
-        binding.simpleButton.setOnClickListener(v -> {
-            geoDataRequester.requestGeoPoly(prompt);
-        });
+        binding.simpleButton.setOnClickListener(v -> geoDataRequester.requestGeoPoly(prompt));
 
         String stringAnswer = GeoWidgetUtils.getGeoPolyAnswerToDisplay(prompt.getAnswerText());
         binding.geoAnswerText.setText(stringAnswer);
@@ -62,15 +57,15 @@ public class GeoTraceWidget extends QuestionWidget {
 
         if (getFormEntryPrompt().isReadOnly()) {
             if (dataAvailable) {
-                binding.simpleButton.setText(org.odk.collect.strings.R.string.view_line);
+                binding.simpleButton.setText(org.odk.collect.strings.R.string.view_polygon);
             } else {
                 binding.simpleButton.setVisibility(View.GONE);
             }
         } else {
             if (dataAvailable) {
-                binding.simpleButton.setText(org.odk.collect.strings.R.string.view_or_change_line);
+                binding.simpleButton.setText(org.odk.collect.strings.R.string.view_or_change_polygon);
             } else {
-                binding.simpleButton.setText(org.odk.collect.strings.R.string.get_line);
+                binding.simpleButton.setText(org.odk.collect.strings.R.string.get_polygon);
             }
         }
 
@@ -83,13 +78,13 @@ public class GeoTraceWidget extends QuestionWidget {
     }
 
     @Override
+    public void clearAnswer() {}
+
+    @Override
     public void setOnLongClickListener(OnLongClickListener l) {
         binding.simpleButton.setOnLongClickListener(l);
         binding.geoAnswerText.setOnLongClickListener(l);
     }
-
-    @Override
-    public void clearAnswer() {}
 
     @Override
     public void cancelLongPress() {
