+++ collect_app/src/test/java/org/odk/collect/android/widgets/geo/GeoPolyDialogFragmentTest.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.widgets.utilities
+package org.odk.collect.android.widgets.geo
 
 import android.R
 import android.os.Bundle
@@ -31,8 +31,6 @@ import org.odk.collect.android.javarosawrapper.SuccessValidationResult
 import org.odk.collect.android.javarosawrapper.ValidationResult
 import org.odk.collect.android.support.CollectHelpers
 import org.odk.collect.android.support.MockFormEntryPromptBuilder
-import org.odk.collect.android.widgets.geo.GeoPolyDialogFragment
-import org.odk.collect.android.widgets.geo.ReferenceGeometryMappableData
 import org.odk.collect.android.widgets.items.GeoSelectChoiceElements
 import org.odk.collect.android.widgets.support.FormElementFixtures.selectChoice
 import org.odk.collect.android.widgets.support.FormElementFixtures.treeElement
